+++ src/main/java/roidrole/patternbanners/recipe/_Recipe.java
@@ -23,23 +23,26 @@ public class _Recipe {
     public static List<BannerPattern> PATTERNS_ONLY_SHAPE = new ArrayList<>();
 
     public static void init(){
-        for (ConfigCategory mapping : ConfigMapping.mappings){
-            ResourceLocation name = new ResourceLocation(Tags.MOD_ID, "recipes/pattern_create/" + mapping.get("hash").getString());
-            if(mapping.containsKey("item") && !mapping.get("item").getString().isEmpty()){
-                ResourceLocation group = new ResourceLocation("");
-                ItemStack patternStack = new ItemStack(pattern, 1, mapping.get("meta").getInt());
-                GameRegistry.addShapelessRecipe(name, group, patternStack,
-                        Ingredient.fromItem(Items.PAPER),
-                        Ingredient.fromStacks(Utils.getItemStack(mapping))
-                );
-            } else if (mapping.containsKey("shap") && mapping.get("shap").getString().length() == 11) {
-                PatternFromShape nextOne = new PatternFromShape(mapping);
-                if(ConfigGeneral.Recipes.craftingTable) {
-                    GameData.register_impl(nextOne.setRegistryName(name));
+        if(ConfigGeneral.Patterns.craftable){
+            for (ConfigCategory mapping : ConfigMapping.mappings){
+                ResourceLocation name = new ResourceLocation(Tags.MOD_ID, "recipes/pattern_create/" + mapping.get("hash").getString());
+                if(mapping.containsKey("item") && !mapping.get("item").getString().isEmpty()){
+                    ResourceLocation group = new ResourceLocation("");
+                    ItemStack patternStack = new ItemStack(pattern, 1, mapping.get("meta").getInt());
+                    GameRegistry.addShapelessRecipe(name, group, patternStack,
+                            Ingredient.fromItem(Items.PAPER),
+                            Ingredient.fromStacks(Utils.getItemStack(mapping))
+                    );
+                } else if (mapping.containsKey("shap") && mapping.get("shap").getString().length() == 11) {
+                    PatternFromShape nextOne = new PatternFromShape(mapping);
+                    if(ConfigGeneral.Recipes.craftingTable) {
+                        GameData.register_impl(nextOne.setRegistryName(name));
+                    }
                 }
             }
         }
-        if(!ConfigGeneral.shapes_pattern){
+
+        if(!ConfigGeneral.Patterns.shapes_pattern){
             for(BannerPattern pattern : BannerPattern.values()){
                 if(pattern.hasPatternItem() || !pattern.hasPattern()){continue;}
                 PATTERNS_ONLY_SHAPE.add(pattern);