+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Unreleased :
 **Added** : 
 - Loom
 - Config to remove crafting table and/or loom recipes
+- Config to fall back to old pattern texture
+- Config to always use fallback texture
 
 **Fixed** : 
 - entryName of injected entry in Bastion